MALLOK HOOKS: minor refactoring of bistro infra
- re-implemented ucs_clear_cache call
- added comments
