+++ src/ucm/bistro/bistro_aarch64.h
@@ -11,10 +11,18 @@
 #include <stdint.h>
 
 #include <ucs/type/status.h>
+#include <ucs/sys/compiler_def.h>
 
 #define UCM_BISTRO_PROLOGUE
 #define UCM_BISTRO_EPILOGUE
 
+typedef struct ucm_bistro_patch {
+    uint32_t reg3;  /* movz    x15, addr, lsl #48 */
+    uint32_t reg2;  /* movk    x15, addr, lsl #32 */
+    uint32_t reg1;  /* movk    x15, addr, lsl #16 */
+    uint32_t reg0;  /* movk    x15, addr          */
+    uint32_t br;    /* br      x15                */
+} UCS_S_PACKED ucm_bistro_patch_t;
 
 /**
  * Set library function call hook using Binary Instrumentation
