+++ src/ucs/arch/cpu.h
@@ -69,4 +69,22 @@ typedef enum ucs_cpu_flag {
 #define UCS_SYS_CACHE_LINE_SIZE    UCS_ARCH_CACHE_LINE_SIZE
 #endif
 
+/**
+ * Clear processor data and instruction caches, intended for
+ * self-modifying code.
+ *
+ * @start start of region to clear cache, including address
+ * @end   end of region to clear cache, excluding address
+ */
+static inline void ucs_clear_cache(void *start, void *end)
+{
+#if HAVE___CLEAR_CACHE
+    /* do not allow global declaration of compiler intrinsic */
+    void __clear_cache(void* beg, void* end);
+
+    __clear_cache(start, end);
+#else
+    ucs_arch_clear_cache(start, end);
+#endif
+}
 #endif
