+++ src/ucs/arch/ppc64/cpu.h
@@ -57,17 +57,12 @@ double ucs_arch_get_clocks_per_sec();
 
 #define ucs_arch_wait_mem ucs_arch_generic_wait_mem
 
-static inline void ucs_clear_cache(void *start, void *end)
+#if !HAVE___CLEAR_CACHE
+static inline void ucs_arch_clear_cache(void *start, void *end)
 {
-#if HAVE___CLEAR_CACHE
-    /* do not allow global declaration of compiler intrinsic */
-    void __clear_cache(void* beg, void* end);
-
-    __clear_cache(start, end);
-#else
     ucs_memory_cpu_fence();
-#endif
 }
+#endif
 
 END_C_DECLS
